requirements.txt
google-cloud-pubsub==2.18.4
google-cloud-resource-manager==1.10.3
google-cloud-storage==2.10.0